+++ drift/lib/src/web/wasm_setup.dart
@@ -259,6 +259,7 @@ final class _ProbeResult implements WasmProbeResult {
           enableMigrations: enableMigrations,
           protocolVersion: sharedWorker!.version,
           newSerialization: sharedWorker.version >= ProtocolVersion.v3,
+          clientLock: await holdClientLock(),
         );
 
         message.sendTo(sharedWorker.send);
@@ -274,6 +275,7 @@ final class _ProbeResult implements WasmProbeResult {
             enableMigrations: enableMigrations,
             protocolVersion: dedicatedWorker.version,
             newSerialization: dedicatedWorker.version >= ProtocolVersion.v3,
+            clientLock: await holdClientLock(),
           );
 
           message.sendTo(dedicatedWorker.send);